+++ changelog.md
@@ -1,4 +1,11 @@
 # Changelog
 
-**./src/file-prep/folder-intl.js**
-* Added missing comma to exports.
+**./src/common/database/row-data/data-case-documents.js**
+* Moved the 'prepareCaseDocumentLinkRows' function to '../db-select.js'
+* This file is now empty.
+
+---
+
+**./src/common/database/db-select.js**
+* Added 'prepareCaseDocumentLinkRows' function from 'dataCaseDocuments'
+* Replaced 'dataCaseDocuments.prepareCaseDocumentLinks' with 'prepareCaseDocumentLinkRows'